EarthSciMLData

Documentation for EarthSciMLData.

EarthSciMLData.DataArray

EarthSciMLData.DataFrequencyInfo

EarthSciMLData.DataSetInterpolator

EarthSciMLData.FileSet

EarthSciMLData.GEOSFP

EarthSciMLData.GEOSFPFileSet

EarthSciMLData.centerpoint_index

EarthSciMLData.description

EarthSciMLData.dimnames

EarthSciMLData.endpoints

EarthSciMLData.interp!

EarthSciMLData.load_interpolator

EarthSciMLData.loadslice

EarthSciMLData.localpath

EarthSciMLData.maybedownload

EarthSciMLData.prune!

EarthSciMLData.relpath

EarthSciMLData.units

EarthSciMLData.url

EarthSciMLData.varnames

EarthSciMLData.DataArray — Type

An array of data.

data: The data.

units: Physical units of the data, e.g. m s⁻¹.

description: Description of the data.

dimnames: Dimensions of the data, e.g. (lat, lon, layer).

EarthSciMLData.DataFrequencyInfo — Type

Information about the temporal frequency of archived data.

start: Beginning of time of the time series.

frequency: Interval between each record.

centerpoints: Time representing the temporal center of each record.

EarthSciMLData.DataSetInterpolator — Type

DataSetInterpolators are used to interpolate data from a FileSet to represent a given time and location. Data is loaded (and downloaded) lazily, so the first time you use it on a for a given dataset and time period it may take a while to load. Each time step is downloaded and loaded as it is needed during the simulation and cached on the hard drive at the path specified by the \$ESMLDATADIR environment variable. The interpolator will also cache data in memory representing the data records for the times immediately before and after the current time step.

EarthSciMLData.FileSet — Type

An interface for types describing a dataset, potentially comprised of multiple files.

To satisfy this interface, a type must implement the following methods:

relpath(fs::FileSet, t::DateTime)

url(fs::FileSet, t::DateTime)

localpath(fs::GEOSFPFileSet, t::DateTime)

DataFrequencyInfo(fs::GEOSFPFileSet, t::DateTime)::DataFrequencyInfo

loadslice(fs::GEOSFPFileSet, t::DateTime, varname)::DataArray

load_interpolator(fs::GEOSFPFileSet, t::DateTime, varname)

varnames(fs::GEOSFPFileSet, t::DateTime)

EarthSciMLData.GEOSFP — Type

A data loader for GEOS-FP data as archived for use with GEOS-Chem classic.

Domain options (as of 2022-01-30):

4x5

0.125x0.15625_AS

0.125x0.15625_EU

0.125x0.15625_NA

0.25x0.3125

0.25x0.3125_AF

0.25x0.3125_AS

0.25x0.3125_CH

0.25x0.3125_EU

0.25x0.3125_ME

0.25x0.3125_NA

0.25x0.3125_OC

0.25x0.3125_RU

0.25x0.3125_SA

0.5x0.625

0.5x0.625_AS

0.5x0.625_CH

0.5x0.625_EU

0.5x0.15625_NA

2x2.5

4x5

C180

C720

NATIVE

c720

See http://geoschemdata.wustl.edu/ExtData/ for current options.

EarthSciMLData.GEOSFPFileSet — Type

GEOS-FP data as archived for use with GEOS-Chem classic.

Domain options (as of 2022-01-30):

4x5

0.125x0.15625_AS

0.125x0.15625_EU

0.125x0.15625_NA

0.25x0.3125

0.25x0.3125_AF

0.25x0.3125_AS

0.25x0.3125_CH

0.25x0.3125_EU

0.25x0.3125_ME

0.25x0.3125_NA

0.25x0.3125_OC

0.25x0.3125_RU

0.25x0.3125_SA

0.5x0.625

0.5x0.625_AS

0.5x0.625_CH

0.5x0.625_EU

0.5x0.15625_NA

2x2.5

4x5

C180

C720

NATIVE

c720

Possible filetypes are:

:A1

:A3cld

:A3dyn

:A3mstC

:A3mstE

:I3

See http://geoschemdata.wustl.edu/ExtData/ for current options.

EarthSciMLData.centerpoint_index — Method

Return the index of the centerpoint closest to the given time.

EarthSciMLData.description — Method

description(itp, t)

Return the description of the data associated with this interpolator.

EarthSciMLData.dimnames — Method

dimnames(itp, t)

Return the dimension names associated with this interpolator.

EarthSciMLData.endpoints — Method

Return the time endpoints correcponding to each centerpoint

EarthSciMLData.interp! — Method

interp!(itp, t, locs)

Return the value of the given variable from the given dataset at the given time and location.

EarthSciMLData.load_interpolator — Method

load_interpolator(fs, t, varname)

Load the data for the given DateTime and variable name as an interpolator from Interpolations.jl.

EarthSciMLData.loadslice — Method

loadslice(fs, t, varname)

Load the data for the given variable name at the given time.

EarthSciMLData.localpath — Method

localpath(fs, t)

Return the local path for the GEOS-FP file for the given DateTime.

EarthSciMLData.maybedownload — Method

maybedownload(fs, t)

Check if the specified file exists locally. If not, download it.

EarthSciMLData.prune! — Method

prune!(pde_sys, prefix)

Remove equations from a PDESystem where a variable in the LHS contains the given prefix but none of the equations have an RHS containing that variable. This can be used to remove data loading equations that are not used in the final model.

EarthSciMLData.relpath — Method

relpath(fs, t)

File path on the server relative to the host root; also path on local disk relative to ENV["ESMLDATADIR"].

EarthSciMLData.units — Method

units(itp, t)

Return the units of the data associated with this interpolator.

EarthSciMLData.url — Method

url(fs, t)

Return the URL for the GEOS-FP file for the given DateTime.

EarthSciMLData.varnames — Method

varnames(fs, t)

Return the variable names associated with this FileSet.